@@ -1,44 +1,46 @@
 {
   "name": "Access2Justice.Web",
   "version": "0.0.0",
+  "private": true,
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --extract-css",
-    "build": "ng build --extract-css",
+    "start": "ng serve",
+    "build": "ng build --extractCss",
     "build:ssr": "npm run build -- --app=ssr --output-hashing=media",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
-  "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.11",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/platform-server": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/animations": "7.2.3",
+    "@angular/common": "7.2.3",
+    "@angular/compiler": "7.2.3",
+    "@angular/core": "7.2.3",
+    "@angular/forms": "7.2.3",
+    "@angular/platform-browser": "7.2.3",
+    "@angular/platform-browser-dynamic": "7.2.3",
+    "@angular/router": "7.2.3",
     "@azure/msal-angular": "^0.1.2",
-    "@nguniversal/module-map-ngfactory-loader": "^5.0.0-beta.5",
     "aspnet-prerendering": "^3.0.1",
     "bootstrap": "^4.1.3",
     "core-js": "^2.4.1",
+    "jquery": "^3.3.1",
     "jspdf": "^1.4.1",
-    "ngx-bootstrap": "^2.0.4",
-    "ngx-spinner": "^1.2.0",
-    "ngx-toastr": "^8.10.0",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "ngx-bootstrap": "^3.2.0",
+    "ngx-spinner": "^7.0.0",
+    "ngx-toastr": "^9.1.1",
+    "popper.js": "^1.14.6",
+    "rxjs": "^6.4.0",
+    "rxjs-compat": "^6.4.0",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "7.3.0",
+    "@angular/compiler-cli": "7.2.3",
+    "@angular/language-service": "7.2.3",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/jspdf": "^1.1.31",
@@ -46,7 +48,7 @@
     "codelyzer": "^4.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "^3.1.4",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
@@ -54,6 +56,6 @@
     "protractor": "^5.4.1",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "3.2.4"
   }
 }
