@@ -1,8 +1,9 @@
+
 import { Injectable } from "@angular/core";
 import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot } from "@angular/router";
-import { Subject } from "rxjs";
-import { Observable } from "rxjs/Observable";
-import { of } from "rxjs/observable/of";
+import { Subject, Observable, of } from "rxjs";
+import {map} from 'rxjs/operators';
+
 import { Global } from "../../global";
 import { LoginService } from "../../shared/login/login.service";
 
@@ -20,10 +21,10 @@ export class AdminAuthGuard implements CanActivate {
     next: ActivatedRouteSnapshot,
     state: RouterStateSnapshot
   ): Observable<any> {
-    let url: string = state.url;
+    const url: string = state.url;
 
     if (this.global.roleInformation) {
-      let findAdmin = this.global.roleInformation.find(role => {
+      const findAdmin = this.global.roleInformation.find(role => {
         return this.adminRoles.includes(role.roleName);
       });
       if (!findAdmin) {
@@ -32,12 +33,12 @@ export class AdminAuthGuard implements CanActivate {
         return of(true);
       }
     } else {
-      return this.checkAdminStatus(url, this.adminRoles).map(value => {
+      return this.checkAdminStatus(url, this.adminRoles).pipe(map(value => {
         if (!value) {
           this.router.navigate(["/401"]);
         }
         return value;
-      });
+      }));
     }
   }
 
@@ -47,7 +48,7 @@ export class AdminAuthGuard implements CanActivate {
   ): Observable<any> {
     let url: string = state.url;
     if (this.global.roleInformation) {
-      let findPortalAdmin = this.global.roleInformation.find(role => {
+      const findPortalAdmin = this.global.roleInformation.find(role => {
         return role.roleName === "PortalAdmin";
       });
       if (!findPortalAdmin) {
@@ -56,12 +57,12 @@ export class AdminAuthGuard implements CanActivate {
         return of(true);
       }
     } else {
-      return this.checkAdminStatus(url, "PortalAdmin").map(value => {
+      return this.checkAdminStatus(url, "PortalAdmin").pipe(map(value => {
         if (!value) {
           this.router.navigate(["/401"]);
         }
         return value;
-      });
+      }));
     }
   }
 
