@@ -1,6 +1,7 @@
 import { HttpClient, HttpHeaders } from "@angular/common/http";
 import { Injectable } from "@angular/core";
-import { Observable } from "rxjs/Observable";
+import { Observable } from "rxjs";
+
 import { api } from "../../../api/api";
 import { Answer } from "./answers";
 import { Question } from "./question";
