@@ -1,5 +1,6 @@
 import { TestBed } from "@angular/core/testing";
-import { Observable } from "rxjs/Rx";
+import { Observable } from "rxjs";
+
 import { Global } from "../../global";
 import { MapLocation } from "../../shared/map/map";
 import { TopicService } from "./topic.service";
