@@ -3,7 +3,8 @@ import { NO_ERRORS_SCHEMA } from "@angular/core";
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 import { FormsModule, NgForm } from "@angular/forms";
 import { Router } from "@angular/router";
-import { of } from "rxjs/observable/of";
+import { of } from "rxjs";
+
 import { Global } from "../global";
 import { MapService } from "../shared/map/map.service";
 import { SearchService } from "../shared/search/search.service";
