@@ -3,7 +3,8 @@ import { HttpClientModule } from "@angular/common/http";
 import { CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA } from "@angular/core";
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 import { RouterModule } from "@angular/router";
-import { of } from "rxjs/observable/of";
+import { of } from "rxjs";
+
 import { Global } from "../global";
 import { MapService } from "../shared/map/map.service";
 import { PaginationService } from "../shared/pagination/pagination.service";
